refactor: move decoder function to new module, remove eval
